% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_soi.R
\name{get_soi}
\alias{get_soi}
\title{Get IRS SOI Tax Stats - County Data}
\source{
https://www.irs.gov/statistics/soi-tax-stats-county-data
}
\usage{
get_soi(year, state, msa = FALSE)
}
\arguments{
\item{year}{the year}

\item{state}{of interest}
}
\description{
A function to download tax stats
}
